:recycle: findpaths: Move tests to `tests.unit.templates`
