+++ tests/unit/services/test_create.py
@@ -1,32 +1,1 @@
 """Unit tests for cutty.services.create."""
-from cutty.filesystems.adapters.dict import DictFilesystem
-from cutty.filesystems.domain.path import Path
-from cutty.templates.adapters.cookiecutter.config import findpaths
-from cutty.templates.domain.config import Config
-
-
-def test_findpaths_template_directory() -> None:
-    """It returns the template directory."""
-    config = Config({}, ())
-    filesystem = DictFilesystem({"{{ cookiecutter.project }}": {}})
-    path = Path(filesystem=filesystem)
-    paths = findpaths(path, config)
-
-    assert next(paths) == path / "{{ cookiecutter.project }}"
-    assert next(paths, None) is None
-
-
-def test_findpaths_other_directories() -> None:
-    """It ignores other directories."""
-    config = Config({}, ())
-    filesystem = DictFilesystem(
-        {
-            "hooks": {"pre_gen_project.py~": ""},
-            "{{ cookiecutter.project }}": {},
-        }
-    )
-    path = Path(filesystem=filesystem)
-    paths = findpaths(path, config)
-
-    assert next(paths) == path / "{{ cookiecutter.project }}"
-    assert next(paths, None) is None
