+++ tests/unit/templates/domain/test_config.py
@@ -2,6 +2,7 @@
 from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path
 from cutty.templates.adapters.cookiecutter.config import findhooks
+from cutty.templates.adapters.cookiecutter.config import findpaths
 from cutty.templates.domain.config import Config
 from cutty.templates.domain.variables import Variable
 
@@ -9,6 +10,33 @@ from cutty.templates.domain.variables import Variable
 def test_config(variable: Variable) -> None:
     """It can store settings and variables."""
     Config(settings={"a-setting": "value"}, variables=(variable,))
+
+
+def test_findpaths_template_directory() -> None:
+    """It returns the template directory."""
+    config = Config({}, ())
+    filesystem = DictFilesystem({"{{ cookiecutter.project }}": {}})
+    path = Path(filesystem=filesystem)
+    paths = findpaths(path, config)
+
+    assert next(paths) == path / "{{ cookiecutter.project }}"
+    assert next(paths, None) is None
+
+
+def test_findpaths_other_directories() -> None:
+    """It ignores other directories."""
+    config = Config({}, ())
+    filesystem = DictFilesystem(
+        {
+            "hooks": {"pre_gen_project.py~": ""},
+            "{{ cookiecutter.project }}": {},
+        }
+    )
+    path = Path(filesystem=filesystem)
+    paths = findpaths(path, config)
+
+    assert next(paths) == path / "{{ cookiecutter.project }}"
+    assert next(paths, None) is None
 
 
 def test_findhooks_none() -> None:
